projects/mydungeon/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
pillow==10.1.0
pandas==2.1.3
python-dotenv==1.0.0
reportlab==4.0.7
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
